#41 updated code according to the peer code review and PR suggestions
