+++ server/src/__mocks__/mockDonatedItem.ts
@@ -1,0 +1,14 @@
+export const updateItem = {
+    itemType: 'Updated Book',
+    currentStatus: 'Received',
+    programId: 1,
+    donorId: 1,
+    dateDonated: new Date().toISOString()
+};
+export const newItem = {
+    itemType: 'Book',
+    currentStatus: 'Received',
+    programId: 1,
+    donorId: 1,
+    dateDonated: new Date().toISOString()
+};
