--- server/src/__mocks__/mockNewItem.ts
@@ -1,7 +1,0 @@
-export const newItem = {
-    itemType: 'Book',
-    currentStatus: 'Received',
-    programId: 1,
-    donorId: 1,
-    dateDonated: new Date().toISOString()
-};
