--- server/src/__mocks__/mockUpdateData.ts
@@ -1,7 +1,0 @@
-export const updateData = {
-    itemType: 'Updated Book',
-    currentStatus: 'Received',
-    programId: 1,
-    donorId: 1,
-    dateDonated: new Date().toISOString()
-};
